Refactor Pacman
