+++ app/src/main/java/dev/baseio/composeplayground/ui/animations/pacman/Pacman.kt
@@ -1,4 +1,4 @@
-package dev.baseio.composeplayground.ui.animations
+package dev.baseio.composeplayground.ui.animations.pacman
 
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.RepeatMode
@@ -23,6 +23,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.graphics.drawscope.Fill
 import dev.baseio.composeplayground.contributors.PushpalRoy
+import dev.baseio.composeplayground.ui.animations.offGray
 
 @Composable
 fun Pacman(modifier: Modifier) {
